Add DbType

+++ test-data-generator/src/main/java/com/asksunny/codegen/CodeGenerator.java
@@ -10,31 +10,28 @@ import com.asksunny.schema.Schema;
 
 public abstract class CodeGenerator {
 
-	protected CodeGenConfig configuration;	
+	protected CodeGenConfig configuration;
 	protected Entity entity;
 	protected Schema schema;
+
 	public CodeGenerator(CodeGenConfig configuration, Entity entity) {
 		super();
 		this.configuration = configuration;
 		this.entity = entity;
 	}
+
 	public CodeGenerator(CodeGenConfig configuration, Schema schema) {
 		super();
 		this.configuration = configuration;
 		this.schema = schema;
 	}
-	
-	
-	
-	
-	protected void writeCode(CodeGenConfig config, File dir, String fileName,
-			String code) throws IOException {
+
+	protected void writeCode(File dir, String fileName, String code) throws IOException {
+
 		File fj = new File(dir, fileName);
-		if (configuration.getOverwriteStrategy() == CodeOverwriteStrategy.IGNORE
-				&& fj.exists()) {
+		if (configuration.getOverwriteStrategy() == CodeOverwriteStrategy.IGNORE && fj.exists()) {
 			return;
-		} else if (configuration.getOverwriteStrategy() == CodeOverwriteStrategy.RENAME_EXISTING
-				&& fj.exists()) {
+		} else if (configuration.getOverwriteStrategy() == CodeOverwriteStrategy.RENAME_EXISTING && fj.exists()) {
 			File newFile = null;
 			for (int i = 1; i < Integer.MAX_VALUE; i++) {
 				newFile = new File(dir, String.format("%s.%03d", fileName, i));
@@ -42,9 +39,11 @@ public abstract class CodeGenerator {
 					break;
 				}
 			}
-			if(!fj.renameTo(newFile)){
+			if (!fj.renameTo(newFile)) {
 				throw new IOException("Failed to rename existing file");
-			}			
+			}
+		} else if (configuration.getOverwriteStrategy() == CodeOverwriteStrategy.OVERWRITE && fj.exists()) {
+			//
 		} else if (fj.exists()) {
 			for (int i = 1; i < Integer.MAX_VALUE; i++) {
 				fj = new File(dir, String.format("%s.%03d", fileName, i));
@@ -53,6 +52,11 @@ public abstract class CodeGenerator {
 				}
 			}
 		}
+
+		if (fj.getParentFile().exists() == false && fj.getParentFile().mkdirs() == false) {
+			throw new IOException("Permission denied to created directory:" + fj.getParentFile().toString());
+		}
+
 		FileWriter fw = new FileWriter(fj);
 		try {
 			fw.write(code);
